MONGOID-4544 Define #touch on all documents (#4521)
