+++ lib/mongoid/association.rb
@@ -3,7 +3,6 @@ require 'mongoid/association/builders'
 require 'mongoid/association/bindable'
 require 'mongoid/association/depending'
 require 'mongoid/association/proxy'
-require 'mongoid/association/touchable'
 
 require 'mongoid/association/many'
 require 'mongoid/association/one'
