+++ spec/app/models/updatable.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Updatable
+  include Mongoid::Document
+
+  field :updated_at, type: BSON::Timestamp
+end
