+++ spec/mongoid/association/referenced/belongs_to_spec.rb
@@ -449,7 +449,7 @@ describe Mongoid::Association::Referenced::BelongsTo do
         end
 
         it 'sets up touch' do
-          expect(Mongoid:: Association::Touchable).to receive(:define_touchable!).with(association)
+          expect(Mongoid::Touchable).to receive(:define_touchable!).with(association)
           association.setup!
         end
       end
@@ -463,7 +463,7 @@ describe Mongoid::Association::Referenced::BelongsTo do
         end
 
         it 'does not set up touch' do
-          expect(Mongoid:: Association::Touchable).not_to receive(:define_touchable!).with(association)
+          expect(Mongoid::Touchable).not_to receive(:define_touchable!).with(association)
           association.setup!
         end
       end
@@ -477,7 +477,7 @@ describe Mongoid::Association::Referenced::BelongsTo do
         end
 
         it 'does not set up touch' do
-          expect(Mongoid:: Association::Touchable).not_to receive(:define_touchable!).with(association)
+          expect(Mongoid::Touchable).not_to receive(:define_touchable!).with(association)
           association.setup!
         end
       end
